requirements.txt
playwright
pytest-playwright
pytest
allure-pytest